feedback: common data use chips, table alignment, restore :focus-visible
